認証関連のディレクトリ構造の変更
