+++ frontend/src/app/(auth)/_components/organisms/BaseLayout/index.tsx
