+++ frontend/src/app/(auth)/_components/organisms/SignUpForm/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState } from 'react';
-import { PhaseType } from '../../../types';
+import { PhaseType } from '../../../sign_up/types';
 import { SignUpConfirmation } from '../SignUpConfirmation';
 import { SignUpInput } from '../SignUpInput';
 import { SignUpThanks } from '../SignUpThanks';
