+++ frontend/src/app/(auth)/_components/organisms/SignUpInput/index.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react';
 import { postValidateSignUp } from '../../../_actions/postValidateSignUp';
 import { useSignUpContext } from '../../../_contexts/SignUpContext';
-import { PhaseType } from '../../../types';
+import { PhaseType } from '../../../sign_up/types';
 import { BaseLayout } from '../BaseLayout';
 
 type Props = {
