+++ frontend/src/app/(auth)/_components/organisms/SignUpThanks/index.test.tsx
