+++ frontend/src/app/(auth)/_components/templates/SignUpTemplate/index.tsx
