+++ frontend/src/app/(auth)/sign_up/page.tsx
@@ -1,4 +1,4 @@
-import { SignUpTemplate } from './_components/templates/SignUpTemplate';
+import { SignUpTemplate } from '../_components/templates/SignUpTemplate';
 
 export default async function SignUpPage() {
   return (
